+++ map/src/render/renderer.ts
@@ -874,6 +874,8 @@ export class MapRendererContent {
         tile_ecef_center_l: [0, 0, 0, 0],
         cam_ecef_center_h: [0, 0, 0, 0],
         cam_ecef_center_l: [0, 0, 0, 0],
+        tile_origin_merc_hl: [0, 0, 0, 0],
+        cam_merc_center_hl: [0, 0, 0, 0],
         light_dir_ecef: [0, 0, 0, 0],
         globe_eye: [ge[0], ge[1], ge[2], ge[3]],
       })